Extract primary action resolution
